chore: implement remaining test cases
